The great asyncing!

+++ commands/deltag.js
@@ -1,14 +1,15 @@
 const sql = require('sqlite');
-exports.run = (client, message, args) => {
+sql.open('./tagsbot.sqlite');
+exports.run = async(client, message, args) => {
   let name = args.join(' ');
-  sql.open('./tagsbot.sqlite').then(() => {
-    return sql.run('DELETE FROM tags WHERE name = ?', name);
-  }).then(() => {
-    return message.channel.sendMessage(`The tag **${name}** has been deleted`);
-  }).then(response => {
-    return response.delete(5000);
-  }).catch(console.error);
-
+  try {
+    const row = await sql.get(`SELECT * FROM tags WHERE name = '${name}'`);
+    if (!row) return message.channel.sendMessage(`A tag with the name **${name}** could not be found.`);
+    await sql.run('DELETE FROM tags WHERE name = ?', name);
+    await message.channel.sendMessage(`The tag **${name}** has been deleted`);
+  } catch (error) {
+    console.error;
+  }
 };
 
 exports.conf = {
